+++ src/utils/functions.ts
@@ -1,16 +1,14 @@
+/* eslint-disable no-useless-catch */
 import api from "@/utils/services/api";
-import {
-  logout,
-  updateUserProfile
-} from "@/utils";
-import { ILogOut, IProfileModifyResponse } from '@/types/responses';
-import { UserProfile } from '@/types/user';
+import { logout, updateUserProfile } from "@/utils";
+import { ILogOut, IProfileModifyResponse } from "@/types/responses";
+import { UserProfile } from "@/types/user";
 
 export const logOut = async (token: string) => {
   try {
     api.resource = logout;
 
-    const res = await api.post<ILogOut>({ body: { token }});
+    const res = await api.post<ILogOut>({ body: { token } });
 
     return res;
   } catch (error) {
@@ -18,10 +16,7 @@ export const logOut = async (token: string) => {
   }
 };
 
-export const editProfile = async (
-  token: string,
-  body: UserProfile
-) => {
+export const editProfile = async (token: string, body: UserProfile) => {
   try {
     api.resource = updateUserProfile;
     api.token = token;
@@ -32,4 +27,15 @@ export const editProfile = async (
   } catch (error) {
     throw error;
   }
+};
+
+export const formatDate = (inputDate: string) => {
+  const date = new Date(inputDate);
+  const day = date.getDate().toString().padStart(2, "0");
+  const month = (date.getMonth() + 1).toString().padStart(2, "0"); // Sumamos 1 al mes, ya que en JavaScript los meses empiezan desde 0 (enero) hasta 11 (diciembre).
+  const year = date.getFullYear().toString();
+  //
+  const formattedDate = `${month}/${day}/${year}`;
+
+  return formattedDate;
 };
