+++ src/views/Home/components/RequiredDocumentsTable/index.tsx
@@ -18,7 +18,7 @@ import ChatIcon from "@mui/icons-material/Chat";
 import MessageModal from "../MessageModal";
 import useMediaQuery from "@mui/material/useMediaQuery";
 import { useTheme } from "@mui/material/styles";
-
+import { formatDate } from "@/utils";
 import RequiredDocumentsTableMobile from "../RequieredDocumentsTableMobile";
 
 interface RequiredDocumentsTableProps {
@@ -32,22 +32,6 @@ const RequiredDocumentsTable = ({
   const isMobile = useMediaQuery(theme.breakpoints.down("md"));
   const [openModal, setOpenModal] = useState(false);
   const [modalMessage, setModalMessage] = useState("");
-
-  // const [componentHeights, setComponentHeights] = useState<number[]>([]);
-
-  function formatDate(inputDate: string) {
-    const date = new Date(inputDate);
-
-    //
-    const day = date.getDate().toString().padStart(2, "0");
-    const month = (date.getMonth() + 1).toString().padStart(2, "0"); // Sumamos 1 al mes, ya que en JavaScript los meses empiezan desde 0 (enero) hasta 11 (diciembre).
-    const year = date.getFullYear().toString();
-
-    //
-    const formattedDate = `${month}/${day}/${year}`;
-
-    return formattedDate;
-  }
 
   if (isMobile) {
     return (
